Reorganize docgen code to make it easier to add new doc outputs
