+++ cmd/docgen/renderers.go
@@ -20,7 +20,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-func renderAssetDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderAssetDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	if len(item.examples) == 0 {
 		return errors.Errorf("no examples found for asset item %s/%s", item.tagValue, item.typeName)
 	}
@@ -55,7 +55,7 @@ func renderAssetDoc(output *strings.Builder, item *documentedItem, session flows
 	return nil
 }
 
-func renderTypeDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderTypeDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	if len(item.examples) == 0 {
 		return errors.Errorf("no examples found for type %s/%s", item.tagValue, item.typeName)
 	}
@@ -79,7 +79,7 @@ func renderTypeDoc(output *strings.Builder, item *documentedItem, session flows.
 	return nil
 }
 
-func renderOperatorDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderOperatorDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	if len(item.examples) == 0 {
 		return errors.Errorf("no examples found for operator %s/%s", item.tagValue, item.typeName)
 	}
@@ -103,7 +103,7 @@ func renderOperatorDoc(output *strings.Builder, item *documentedItem, session fl
 	return nil
 }
 
-func renderContextDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderContextDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	if len(item.examples) == 0 {
 		return errors.Errorf("no examples found for context item %s/%s", item.tagValue, item.typeName)
 	}
@@ -127,7 +127,7 @@ func renderContextDoc(output *strings.Builder, item *documentedItem, session flo
 	return nil
 }
 
-func renderFunctionDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderFunctionDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	if len(item.examples) == 0 {
 		return errors.Errorf("no examples found for function %s", item.tagValue)
 	}
@@ -157,7 +157,7 @@ func renderFunctionDoc(output *strings.Builder, item *documentedItem, session fl
 	return nil
 }
 
-func renderEventDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderEventDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	// try to parse our example
 	exampleJSON := []byte(strings.Join(item.examples, "\n"))
 	event, err := events.ReadEvent(exampleJSON)
@@ -192,7 +192,7 @@ func renderEventDoc(output *strings.Builder, item *documentedItem, session flows
 	return nil
 }
 
-func renderActionDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderActionDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	// try to parse our example
 	exampleJSON := []byte(strings.Join(item.examples, "\n"))
 	action, err := actions.ReadAction(exampleJSON)
@@ -239,7 +239,7 @@ func renderActionDoc(output *strings.Builder, item *documentedItem, session flow
 	return nil
 }
 
-func renderTriggerDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderTriggerDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	// try to parse our example
 	exampleJSON := json.RawMessage(strings.Join(item.examples, "\n"))
 	trigger, err := triggers.ReadTrigger(session.Assets(), exampleJSON, assets.PanicOnMissing)
@@ -270,7 +270,7 @@ func renderTriggerDoc(output *strings.Builder, item *documentedItem, session flo
 	return nil
 }
 
-func renderResumeDoc(output *strings.Builder, item *documentedItem, session flows.Session) error {
+func renderResumeDoc(output *strings.Builder, item *TaggedItem, session flows.Session) error {
 	// try to parse our example
 	exampleJSON := json.RawMessage(strings.Join(item.examples, "\n"))
 	resume, err := resumes.ReadResume(session.Assets(), exampleJSON, assets.PanicOnMissing)
